+++ app/src/main/res/layout/fragment_alert.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="layout.Alert">
+    tools:context=".Alert">
 
 
     <TextView
