+++ app/src/main/res/layout/fragment_wave.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="layout.Wave">
+    tools:context=".Wave">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
